add: modified my script to allow it run multiple dataset

+++ hw4/scripts/matrix.sh
@@ -9,8 +9,6 @@ MAKEFILE_TARGET="Makefile"
 SOURCE_DIR="../src"
 
 DATASET_DIR="../data-set"
-DATASET_TARGET="data2_6"
-DATASET_ANSWER="ans2_6"
 
 # Some setting for synchronize different node
 CURRENT_DIR=$(pwd)
@@ -28,32 +26,52 @@ done
 cp $SOURCE_DIR/$TARGET1 .
 cp $SOURCE_DIR/$TARGET2 .
 cp $SOURCE_DIR/$MAKEFILE_TARGET .
-cp $DATASET_DIR/$DATASET_TARGET .
-cp $DATASET_DIR/$DATASET_ANSWER .
 make
 
-# Create directories in parallel on each host
-parallel-ssh -h hosts -i "mkdir -p ~/$SUB_DIR"
+# Define datasets and their corresponding answers
+# Initialize empty arrays
+DATASETS=()
+ANSWERS=()
+# Generate dataset and answer names
+for i in {1..10}; do
+  DATASETS+=("data1_$i" "data2_$i")
+  ANSWERS+=("ans1_$i" "ans2_$i")
+done
 
-# Copy the executable to each host in parallel
-parallel-scp -h hosts -r matmul ~/$SUB_DIR
+# Loop over each dataset
+for i in "${!DATASETS[@]}"; do
+  DATASET_TARGET=$DATASET_DIR/"${DATASETS[$i]}"
+  DATASET_ANSWER=$DATASET_DIR/"${ANSWERS[$i]}"
+  echo "Run Test $DATASET_TARGET"
 
-# (Optional) Insert commands here that use the copied files on remote hosts
-mpirun -np 4 --hostfile mpi_hosts matmul < $DATASET_TARGET > result.log
+  # Create directories in parallel on each host
+  parallel-ssh -h hosts -i "mkdir -p ~/$SUB_DIR"
+  # Copy the executable to each host in parallel
+  parallel-scp -h hosts -r matmul ~/$SUB_DIR
 
-# Extracting the matrix from the output log
-# Assuming the matrix lines do not contain the word 'MPI'
-grep -v 'MPI' result.log > extracted_matrix
-grep 'MPI' result.log > time.log
+  mpirun -np 4 --hostfile mpi_hosts matmul < $DATASET_TARGET > result.log
 
-# Comparing the extracted matrix with the answer file
-diff extracted_matrix $DATASET_ANSWER
+  # Extracting the matrix from the output log
+  # Assuming the matrix lines do not contain the word 'MPI'
+  grep -v 'MPI' result.log > extracted_matrix
+  grep 'MPI' result.log > time.log
 
-# Clean up files on remote hosts in parallel
-parallel-ssh -h hosts -i "rm -rf ~/$SUB_DIR"
+  # Comparing the extracted matrix with the answer file
+  diff extracted_matrix $DATASET_ANSWER
+
+  # Check if diff returned a non-zero exit status
+  if [ $? -ne 0 ]; then
+    echo "Error: Test $DATASET_TARGET failed. Stopping."
+    exit 1
+  fi
+
+  # Clean up files on remote hosts in parallel
+  parallel-ssh -h hosts -i "rm -rf ~/$SUB_DIR"
+
+done
 
 # Local clean up
-rm $TARGET1 $TARGET2 $MAKEFILE_TARGET $DATASET_TARGET $DATASET_ANSWER
+rm $TARGET1 $TARGET2 $MAKEFILE_TARGET
 rm matmul hosts result.log extracted_matrix
 rm mpi_hosts
 # rm time.log
